pyproject.toml
[tool.poetry]
name = "django-expense-tracker"
version = "0.1.0"
description = ""
authors = ["anshuman265 <verma77712@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
